Refine password reset process
Style 'forgot my password' and 'change password' templates. Allow for emails
to be sent by application. Authentication is kept in a file hidden from the
repository. This hidden file is required by app to load up, and as such, the
app will not load in production in its current state. This cannot be deployed
until a fix is found.